change Datatype

+++ CleanArchitecture.Core/CleanArchitecture.Core/ViewModels/Transaction/GetPairBalanceResponce.cs
@@ -6,10 +6,10 @@ namespace CleanArchitecture.Core.ViewModels.Transaction
 {
     public class GetPairBalanceResponce
     {
-        public string Currency { get; set; }
-        public double Balance { get; set; }
-        public double Available { get; set; }
-        public int Pending { get; set; }
+        public string  Currency { get; set; }
+        public decimal Balance { get; set; }
+        public decimal Available { get; set; }
+        public long Pending { get; set; }
         public string CryptoAddress { get; set; }
         public bool Requested { get; set; }
         public object Uuid { get; set; }
